Cambios funciones
I removed trymcGroup & SNIPS functions. These function were useful to
modify the data for analysis and then they applied function mantelcor
on the modified numbers. The idea of this change is to do the
modification of data in another function and then  use mantelcorGroup
with the new changes as arguments.